:heavy_plus_sign: Add sonar-plugin-api-impl for unit tests :white_check_mark: Add test on plugin
